Create pages view
- Create home.js page.
- Create detail-restaurant.js page.
- Create favorite-restaurants.js page.
- Move parallax-hero, restaurant section, and dish section to home.js.
- Fix styling of main.
