+++ src/scripts/views/pages/detail-restaurant.js
@@ -1,0 +1,11 @@
+const DetailRestaurant = {
+  async render() {
+    return `
+    `;
+  },
+
+  async afterRender() {
+  },
+};
+
+export default DetailRestaurant;
