+++ src/templates/index.html
@@ -25,23 +25,8 @@
         </ul>
       </div>
     </nav>
-    <section class="parallax">
-      <div class="parallax__inner">
-        <h2>Fityan's French Restaurant</h2>
-        <p>Experience French cuisine in a different restaurant scenes</p>
-      </div>
-    </section>
   </header>
-  <main id="main" class="wrapper" tabindex="0">
-    <section class="section">
-      <h2 class="section__title">Explore Restaurant</h2>
-      <restaurant-list class="restaurants"></restaurant-list>
-    </section>
-    <section class="section">
-      <h2 class="section__title">Explore Dish</h2>
-      <dish-list class="dishes"></dish-list>
-    </section>
-  </main>
+  <main id="main" tabindex="0"></main>
   <footer class="footer">
     <div class="wrapper">
       copyright &copy; 2022 by <strong>Fresto</strong>
